Moved LocalRecord expected path, this now matches the online download URL path
